Ported Functions to the Tidyverse Styleguide and Major Updates
The following patterns have been ported:

Candle (Bullish/Bearish) patterns
Engulfing patterns
Harami patterns

Internal Functions have been created
Candle features
Candle-body featues
Shadow features

All functions are collected in seperate wrappers with switch inside.

New Function:

candle_pattern

This function uses NSE, and is a highlevel wrapper to apply multiple patterns to the OHLC data.

New functionality:

The library is basically only dependent on base R and Rlang. The OHLC data is no longer required to be XTS format, and supports all data types.

Updated MSFT:

The MSFT data is now extended to to present date, and is converted to data.frame.

There is a bug, which I cant fix at the moment where xts-objects would be converted to xts-matrices for exported data. This breaks all functionality in the examples.

It is currently not known wether this is a bug, feature or whatever. All that is known is that this is not working as intended.

It will require heavy-testing in the future to check wether this error is persistent. No bugreport or questions are submitted regarding this.

NOTE: For debugging all unused patterns have been removed from the library temporarily. These will be added gradually as the development goes on.