Added remaining patterns, startup message and error-handling
The remaining patterns have been added

1. Kick patterns
2. Tremds
3. Piercing Lines
4. Dark Cloud Cover

A error-handler is on its way, but it needs to be refined. It uses CLI.

The goal of the function is to check wether the passed ohlc is indeed a two dimensional data frame object, and tha it includes OHLC-prices. It also checks wether these are of numeric/integer types.

Added a startup message to the library.